remove duplicates of gov.uk widget classes, separate fields and widgets into appropriately-named modules
